refactor: merge guest dashboards and standard dash
