--- packages/frontend/src/modules/dashboard/pages/guest-dashboard.css
@@ -1,15 +1,0 @@
-.app-link {
-  text-decoration: none;
-  color: inherit;
-}
-
-.app-link:hover {
-  text-decoration: none;
-}
-
-/* Thank you bootstrap, we will take it from here */
-.row > * {
-  margin-bottom: 0.5rem;
-  padding-left: 0.25rem;
-  padding-right: 0.25rem;
-}
